feat: генерация случайной строки не более 20 символов

+++ src/models/UrlModel/UrlModel.ts
@@ -3,6 +3,16 @@ import { Statistics } from "../StatisticsModel/Statistics.ts"
 import type { Request, Response } from "express"
 import type { IUrlDto } from "./dto.ts"
 
+const generateString = (maxLength = 20) => {
+    var abc = "abcdefghijklmnopqrstuvwxyz1234567890";
+        var rs = "";
+        const length = Math.round(Math.random()*maxLength)
+        while (rs.length < length) {
+            rs += abc[Math.floor(Math.random() * abc.length)];
+        }
+        return rs
+}
+
 export const UrlModel = {
     shortenUrl: async (req: Request<any, IUrlDto>, res: Response): Promise<void> => {
         const { originalUrl, alias, expiresAt }: IUrlDto = req.body
@@ -10,8 +20,8 @@ export const UrlModel = {
         const url = new URL(originalUrl)
         const { hostname } = url
 
-        const shortUrl = `${hostname}/${alias}`
-        await Url.create({originalUrl, shortUrl, alias, expiresAt: new Date(), createdAt: new Date()})
+        const shortUrl = `${hostname}/${alias || generateString()}`
+        await Url.create({originalUrl, shortUrl, alias, expiresAt: new Date(expiresAt), createdAt: new Date()})
 
         res.json(shortUrl)
     },
@@ -19,11 +29,13 @@ export const UrlModel = {
         const { shortUrl } = req.params
         const record = await Url.findOne({ where: { shortUrl }})
 
-        if (!record) throw new Error('Ссылка не найдена')
+        if (!record) throw new Error('Ссылка не существует')
 
         const { id } = record.dataValues
         const { originalUrl } = record.dataValues
         const { ip } = req
+
+        console.log(ip)
 
         await Statistics.create({ urlId: id, ip: String(ip), createdAt: new Date() })
         res.send(originalUrl)
@@ -32,7 +44,7 @@ export const UrlModel = {
         const { shortUrl } = req.params
         const record = await Url.findOne({ where: { shortUrl }})
 
-        if (!record) throw new Error('Ссылка не найдена')
+        if (!record) throw new Error('Ссылка не существует')
 
         const { id, originalUrl, createdAt } = record.dataValues
         const clickRecords = await Statistics.findAll({ raw: true, where: { urlId: id }})
@@ -41,6 +53,14 @@ export const UrlModel = {
     },
     deleteUrlInfo: async (req, res): Promise<void> => {
         const { shortUrl } = req.params
+
+        const urlRecord = await Url.findOne({ where: { shortUrl }})
+
+        if (!urlRecord) throw new Error('Ссылка не существует')
+
+        const { id } = urlRecord.dataValues
+
+        await Statistics.destroy({ where: { urlId: id }})
         await Url.destroy({ where: { shortUrl }})
         res.send('Ссылка удалена')
     },
